Change DELETE operation so that a blog can be deleted only by the user who added it. Ex_4.21 BlogList expansion step9

+++ controllers/blogs.js
@@ -76,7 +76,25 @@ blogsRouter.put('/:id', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndDelete(request.params.id)
+
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+  const user = await User.findById(decodedToken.id)
+  console.log(user)
+
+  if (!user) {
+    return response.status(400).json({ error: 'UserId missing or not valid' })
+  }
+  const blogToDelete = await Blog.findById(request.params.id)
+  console.log(blogToDelete)
+  if (user._id.toString() === blogToDelete.user.toString()) {
+    await Blog.findByIdAndDelete(request.params.id)
+  } else {
+    return response.status(401).json({ error: 'userId  invalid' })
+  }
+
   response.status(204).end()
 })
